# bos-workspace/backend-nestjs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy chi tiết 1 người dùng",
        "tags": [
          "Users (Người dùng)"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật thông tin",
        "tags": [
          "Users (Người dùng)"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa tài khoản",
        "tags": [
          "Users (Người dùng)"
        ]
      }
    },
    "/api/v1/signatures": {
      "get": {
        "operationId": "UserSignaturesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy tất cả các mẫu chữ ký/con dấu của tôi",
        "tags": [
          "User Signatures (Mẫu chữ ký & Con dấu cá nhân)"
        ]
      },
      "post": {
        "operationId": "UserSignaturesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSignatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo mới một mẫu chữ ký/con dấu",
        "tags": [
          "User Signatures (Mẫu chữ ký & Con dấu cá nhân)"
        ]
      }
    },
    "/api/v1/signatures/{id}/default": {
      "patch": {
        "operationId": "UserSignaturesController_setDefault",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Thiết lập mẫu chữ ký mặc định",
        "tags": [
          "User Signatures (Mẫu chữ ký & Con dấu cá nhân)"
        ]
      }
    },
    "/api/v1/signatures/{id}": {
      "delete": {
        "operationId": "UserSignaturesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa một mẫu chữ ký",
        "tags": [
          "User Signatures (Mẫu chữ ký & Con dấu cá nhân)"
        ]
      }
    },
...
    "/api/v1/workflow-pipeline/steps": {
      "post": {
        "operationId": "WorkflowStepsController_createStep",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStepDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo bước duyệt mới vào Version DRAFT",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
    "/api/v1/workflow-pipeline/steps/{id}": {
      "patch": {
        "operationId": "WorkflowStepsController_updateStep",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStepDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật bước duyệt",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      },
      "delete": {
        "operationId": "WorkflowStepsController_removeStep",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa bước duyệt",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
    "/api/v1/workflow-pipeline/transitions": {
      "post": {
        "operationId": "WorkflowStepsController_createTransition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tạo đường rẽ nhánh giữa 2 Bước",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
    "/api/v1/workflow-pipeline/transitions/{id}": {
      "patch": {
        "operationId": "WorkflowStepsController_updateTransition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransitionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cập nhật logic rẽ nhánh",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      },
      "delete": {
        "operationId": "WorkflowStepsController_removeTransition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Xóa đường rẽ nhánh",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
    "/api/v1/workflow-pipeline/versions/{versionId}": {
      "get": {
        "operationId": "WorkflowStepsController_getPipeline",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy toàn bộ sơ đồ các bước và rẽ nhánh",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
    "/api/v1/workflow-pipeline/steps/{id}/candidates": {
      "get": {
        "operationId": "WorkflowStepsController_getStepCandidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lấy danh sách nhân sự ứng viên của một bước",
        "tags": [
          "Workflow Pipeline (Vẽ luồng quy trình)"
        ]
      }
    },
...
  },
  "info": {
    "title": "BOS Core Engine API",
    "description": "Tài liệu API cho Nền tảng Quản trị Doanh nghiệp (BOS v2.0)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@bos.com",
            "description": "Email đăng nhập"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn Admin",
            "description": "Họ và tên"
          },
          "departmentId": {
            "type": "number",
            "example": 1,
            "description": "ID Phòng ban"
          },
          "roleId": {
            "type": "number",
            "example": 1,
            "description": "ID Vai trò"
          },
          "tenantId": {
            "type": "number",
            "example": 1,
            "description": "ID Doanh nghiệp (Chỉ Super Admin dùng)"
          }
        }
      },
      "CreateSignatureDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên định danh mẫu chữ ký/con dấu",
            "example": "Chữ ký chính thức"
          },
          "type": {
            "type": "string",
            "description": "Kiểu mẫu chữ ký",
            "enum": [
              "DRAW",
              "IMAGE",
              "STAMP"
            ],
            "example": "DRAW"
          },
          "signatureUrl": {
            "type": "string",
            "description": "Dữ liệu Base64 hình ảnh chữ ký/con dấu",
            "example": "data:image/png;base64,..."
          }
        },
        "required": [
          "name",
          "type",
          "signatureUrl"
        ]
      },
...
      "CreateStepDto": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "number",
            "description": "ID của Workflow Version",
            "example": 2
          },
          "name": {
            "type": "string",
            "description": "Tên bước",
            "example": "Trưởng phòng phê duyệt"
          },
          "stepType": {
            "type": "string",
            "description": "Loại bước (USER_TASK, SYSTEM_TASK)",
            "default": "USER_TASK"
          },
          "permissions": {
            "type": "object",
            "description": "Quyền xem/sửa dữ liệu tại bước này (Step-level RBAC)",
            "example": {
              "total_amount": "READ",
              "reason": "WRITE",
              "note": "HIDDEN"
            }
          },
          "orderIndex": {
            "type": "number",
            "description": "Thứ tự hiển thị trên sơ đồ",
            "default": 0
          }
        },
        "required": [
          "versionId",
          "name"
        ]
      },
      "UpdateStepDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Tên bước",
            "example": "Trưởng phòng phê duyệt"
          },
          "stepType": {
            "type": "string",
            "description": "Loại bước (USER_TASK, SYSTEM_TASK)",
            "default": "USER_TASK"
          },
          "permissions": {
            "type": "object",
            "description": "Quyền xem/sửa dữ liệu tại bước này (Step-level RBAC)",
            "example": {
              "total_amount": "READ",
              "reason": "WRITE",
              "note": "HIDDEN"
            }
          },
          "orderIndex": {
            "type": "number",
            "description": "Thứ tự hiển thị trên sơ đồ",
            "default": 0
          }
        }
      },
      "CreateTransitionDto": {
        "type": "object",
        "properties": {
          "fromStepId": {
            "type": "number",
            "description": "Từ Bước (ID)",
            "example": 1
          },
          "toStepId": {
            "type": "number",
            "description": "Tới Bước (ID)",
            "example": 2
          },
          "conditionLogic": {
            "type": "object",
            "description": "Logic rẽ nhánh (VD: total_amount > 20000000)",
            "example": {
              "field": "total_amount",
              "operator": ">",
              "value": 20000000
            }
          },
          "autoSkip": {
            "type": "boolean",
            "description": "Có tự động bỏ qua nếu thỏa mãn logic không?",
            "default": false
          }
        },
        "required": [
          "fromStepId",
          "toStepId"
        ]
      },
      "UpdateTransitionDto": {
        "type": "object",
        "properties": {
          "conditionLogic": {
            "type": "object",
            "description": "Logic rẽ nhánh (VD: total_amount > 20000000)",
            "example": {
              "field": "total_amount",
              "operator": ">",
              "value": 20000000
            }
          },
          "autoSkip": {
            "type": "boolean",
            "description": "Có tự động bỏ qua nếu thỏa mãn logic không?",
            "default": false
          }
        }
      },
...
    }
  }
}